+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 <!--
-## [Unreleased]
+## [1.0.0]
 ### Added
 ### Changed
 ### Deprecated
@@ -11,6 +11,30 @@
 -->
 
 <!-- ## [v-inc] ${YEAR4}-${MONTHNUMBER}-${DATE} -->
+
+## [1.0.0] 2022-03-30
+### Added
+- New command - `Project Notes: Edit Tags Settings File`
+- New command - `Project Notes: Restore Tag Settings File`
+- All tag settings are now stored in `project-notes-tag-settings.json`
+- Added schema json file for settings file
+- Light and Dark theme aware highlighting
+- New `Tags` can be added to the [settings](SETTINGS.md) file
+- All `Tags` can be enabled or disabled individually in the [settings](SETTINGS.md) file
+- Highlighting of user definable `Tag Blocks`
+- Snippets for the pre-defined tag blocks
+- Added [settings](SETTINGS.md) documentation file
+- Added [Upgrading from previous versions](UPGRADING.md) documentation file
+- New screenshots to reflect the new changes
+### Fixed
+- Highlighting in multiline comments was not working properly
+- Opening project note files not adding lowercase extension files to pick list
+### Changed
+- Now uses text decorations for highlighting
+- File link trigger `File: filename.md` has been changed to `Project File: filename.md`
+- File link trigger now requires the `.MD` extension to remove any confusion on notes file format
+### Removed
+- Dependency on TextMateRules for syntax highlighting
 
 ## [0.0.5] 2022-02-04
 ### Added